+++ Explorer/Assets/DCL/Communities/EventInfo/EventInfoView.cs
@@ -120,7 +120,7 @@ namespace DCL.Communities.EventInfo
                 if (!DateTime.TryParse(eventData.Recurrent_dates[i], null, DateTimeStyles.RoundtripKind, out DateTime date)) continue;
                 if (date < nextStartAt) continue;
 
-                FormatEventString(date, eventData.Duration, eventSchedulesStringBuilder);
+                EventUtilities.FormatEventString(date, eventData.Duration, eventSchedulesStringBuilder);
 
                 if (i < eventData.Recurrent_dates.Length - 1)
                     eventSchedulesStringBuilder.Append("\n");
@@ -129,41 +129,6 @@ namespace DCL.Communities.EventInfo
             var result = eventSchedulesStringBuilder.ToString();
             eventSchedulesStringBuilder.Clear();
             return result;
-        }
-
-        private static void FormatEventString(DateTime utcStart, double durationMs, StringBuilder sb)
-        {
-            TimeSpan duration = TimeSpan.FromMilliseconds(durationMs);
-            DateTime utcEnd = utcStart.Add(duration);
-
-            TimeZoneInfo localZone = TimeZoneInfo.Local;
-            DateTime localStart = TimeZoneInfo.ConvertTimeFromUtc(utcStart, localZone);
-            DateTime localEnd = TimeZoneInfo.ConvertTimeFromUtc(utcEnd, localZone);
-
-            TimeSpan offset = localZone.GetUtcOffset(localStart);
-
-            var day = localStart.ToString("dddd");
-            sb.Append(char.ToUpper(day[0]));
-            sb.Append(day, 1, day.Length - 1);
-            sb.Append(", ");
-
-            var month = localStart.ToString("MMM");
-            sb.Append(char.ToUpper(month[0]));
-            sb.Append(month, 1, month.Length - 1);
-            sb.Append(' ');
-            sb.Append(localStart.ToString("dd"));
-            sb.Append(" from ");
-
-            sb.Append(localStart.ToString("hh:mmtt").ToLowerInvariant());
-            sb.Append(" to ");
-            sb.Append(localEnd.ToString("hh:mmtt").ToLowerInvariant());
-
-            sb.Append(" (UTC");
-            double offsetHours = offset.TotalHours;
-            if (offsetHours >= 0)
-                sb.Append('+');
-            sb.Append(((int)offsetHours).ToString());
-            sb.Append(')');
         }
 
         public void UpdateInterestedButtonState() =>
